+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 
 ### 0.1.12
 
-* Loggers no longer modify the message context (#25)
+* Logger.log() no longer accepts invalid log levels (#26), to match PSR-3.
+* Loggers no longer modify the message context (#25).
 * Removed incorrect `readme` clause in `package.json`.
 
 ### 0.1.11